user_study/voice_memos/requirements.txt
openai-whisper
tqdm
torch 